requirements.txt
streamlit
feedparser
groq
google-generativeai
python-dotenv
requests